Rename module 'section-manager-click-event.js => section-manager-click-handler.js'
