+++ src/handlers/section-manager-click-handler.js
